backend/requirements.txt
flask==2.2.5
flask-cors==4.0.1
requests==2.31.0
python-dotenv==0.21.1
PyPDF2==3.0.1
pdfminer.six==20221105
